Add files via upload
The underground part of the anthill has been added. Ants now know how to pass the maze. Added a new type of ants - the queen. The ants are now emerging from the larvae. There is a new object in the anthill-a food warehouse, where ants can put food, and then eat it. Added small improvements. Fixed small game bugs.